Allow the window to be destroyed and recreated as-needed.
Also fixed a bug in the close callback where a call to
glfwDestroyWindow was made during an event callback. This of course
sent my test code into a swirling death spiral once I started
destroying windows.

Destroying the window is preferable, on the upside, but on the
downside, destruction of the window must be posted to the context so it
happens outside an event callback.