Basic event dispatch framework for windows.
Currently, this does a few things for each event:

For anything which responds to handle_event and is passed an event,
that object can only receive the event once. If it attempts to redirect
the event back to itself via a loop, then the chain is broken to avoid
cycling through targets infinitely.

1. If the event has a target, try to get the target to handle the event.
  If the target changes once it's been handled _and_ the new target
  hasn't already had a chance to handle the event, then it does so.
  Repeat ad nauseum until there are no new targets.

2. Once the target loop ends, being iterating upward through the target's
  superviews, passing the event to any superview that hasn't already
  received the event. Superviews that've already received the event are
  skipped, though this doesn't stop the event from being forwarded
  to all superviews.

3. Then, all leaf view and its superviews in the window's view
  hierarchy are passed the event. This works more or less the same as
  targets.

Events without targets go directly to all leaf views. Avoid target-less
events.

If at any point stop_propagation! is called on the event, the event is
considered handled and nothing else will be able to handle it. Calling
cancel! on an event _does not end event handling_, it merely marks the
event as cancelled. Whether a view respects that is up to the view,
though failing to respect a cancelled event is bad form.