Add basic driver implementation.
The Driver and BufferedDriver classes mainly exist to simplify drawing
2D stuff in GL, so they handle, more or less, the basic stuff when
drawing anything.

Right now, only a draw_quad method is implemented, though that should
be all that's needed for most things.

Text will probably be drawn into a texture and drawn as a quad, but
we'll see -- I'll have to find something that lets me easily access
font data.