Stuff most drawing logic into __draw__.
Ok, after passing out yesterday and forgetting about this, I bothered
checking whether invalidated regions are ever reset for regular views:
answer is no.

So, that's fixed. Stuffed calls to draw and draw_subviews into an
__draw__(driver) method. Ideally, no class other than Window should
override __draw__, but it's permissible as a matter of preparing state
and determining whether a view and its subviews are actually drawn.

Oddly enough, there is no way to actually hide a view yet. That is
apparently the realm of windows and only windows right now. Odd.